Refactor Actions to separate files ♻
