+++ front-end/src/actions/API/restActions.js
@@ -1,0 +1,9 @@
+import { GET_DATA } from '../types';
+
+export const setData = (data) => {
+  return {
+    type: GET_DATA,
+    payload: data,
+  };
+};
+
