+++ front-end/src/actions/API/subscriptionActions.js
@@ -1,0 +1,8 @@
+import { SUBSCRIPTION_MESSAGE } from '../types';
+
+export const addSubMessage = (destination, message) => {
+  return {
+    type: SUBSCRIPTION_MESSAGE,
+    payload: { destination, message },
+  };
+};
