+++ front-end/src/actions/readyActions.js
@@ -1,0 +1,5 @@
+import { PLAYER_READY } from './types';
+
+export const toggleReady = () => {
+  return { type: PLAYER_READY };
+};
